+++ civis/civis.py
@@ -13,6 +13,47 @@ RETRY_CODES = [429, 502, 503, 504]
 
 
 def find(object_list, filter_func=None, **kwargs):
+    """Return the objects from ``object_list`` that satisfy the filters.
+
+    Parameters
+    ----------
+    object_list : iterable
+        An iterable of arbitrary objects, particularly those with
+        attributes that can be targeted by the filters in ``kwargs``.
+        A major use case is a list of ``civis.response.Response`` objects.
+    filter_func : callable, optional
+        A one-argument function. If specified, ``kwargs`` are ignored.
+        An ``object`` from the input iterable is kept in the returned list
+        if and only if ``bool(filter_func(object))`` is ``True``.
+    **kwargs
+        Key-value pairs for more fine-grained filtering; they cannot be used
+        in conjunction with ``filter_func``. All keys must be strings.
+        For an ``object`` from the input iterable to be included in the
+        returned list, all the ``key``s must be attributes of ``object``, plus
+        any one of the following conditions for a given ``key``:
+        - ``value`` is a one-argument function and
+          ``bool(value(getattr(object, key)))`` is ``True``
+        - ``value`` is ``True``
+        - ``getattr(object, key)`` is equal to ``value``
+
+    Returns
+    -------
+    list
+
+    Examples
+    --------
+    >>> import civis
+    >>> client = civis.APIClient()
+    >>> # creds is a list of civis.response.Response objects
+    >>> creds = client.credentials.list()
+    >>> # target_creds contains civis.response.Response objects
+    >>> # with the attribute 'name' == 'username'
+    >>> target_creds = find(creds, name='username')
+
+    See Also
+    --------
+    civis.find_one
+    """
     _func = filter_func
     if not filter_func:
         def default_filter(o):
@@ -35,6 +76,18 @@ def find(object_list, filter_func=None, **kwargs):
 
 
 def find_one(object_list, filter_func=None, **kwargs):
+    """Return one object (or ``None``) from ``civis.find``.
+
+    The arguments are the same as those for ``civis.find``.
+
+    Returns
+    -------
+    object or None
+
+    See Also
+    --------
+    civis.find
+    """
     results = find(object_list, filter_func, **kwargs)
 
     return results[0] if results else None